tests/wpt/tests/html/canvas/offscreen/fill-and-stroke-styles/2d.pattern.paint.norepeat.outside_arc.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.pattern.paint.norepeat.outside_arc</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.pattern.paint.norepeat.outside_arc</h1>
<p class="desc"></p>


<script>
promise_test(async t => {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);
  var response = await fetch('/images/red-16x16.png')
  var blob = await response.blob();
  var img = await createImageBitmap(blob);
  var pattern = ctx.createPattern(img, 'no-repeat');
  ctx.fillStyle = '#0f0';
  ctx.fillRect(0, 0, 100, 50);
  ctx.fillStyle = pattern;
  ctx.beginPath();
  ctx.arc(0, 0, 50, 0, Math.PI * 2);
  ctx.closePath();
  ctx.fill();
  ctx.fillStyle = '#0f0';
  ctx.fillRect(0, 0, 16, 16);
  _assertPixel(canvas, 1,1, 0,255,0,255);
  _assertPixel(canvas, 20,1, 0,255,0,255);
  _assertPixel(canvas, 1,20, 0,255,0,255);
  _assertPixel(canvas, 48,48, 0,255,0,255);

}, "");
</script>
